Made it so that the player shows up behind a door
changed the format of the read map file. Now when it sees a 'P' it looks for two numbers afterwards that say what the previous room the player came from is. It then compares this number to a number stored in the game class. If the room the player was previously in matches this number, that is where it puts the player